# api-docs/api/swagger/swagger.yaml
openapi: 3.0.0
info:
    title: 'Guardian'
    description: 'The Guardian is a modular open-source solution that includes best-in-class identity management and decentralized ledger technology (DLT) libraries. At the heart of the Guardian solution is a sophisticated Policy Workflow Engine (PWE) that enables applications to offer a requirements-based tokenization implementation.'
    version: '1.2.1'
    contact:
        name: 'API developer'
        url: 'https://envisionblockchain.com'
        email: 'info@envisionblockchain.com'
    license:
        name: 'Apache 2.0'
        url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
    - url: /api/v1
      description: version 1.0
components:
    securitySchemes:
        bearerAuth:
            type: 'http'
            scheme: 'bearer'
            bearerFormat: 'JWT'
    schemas:
...
        TrustChains:
            type: object
            required:
                - chain
                - userMap
            properties:
                chain:
                    type: array
                    items:
                        type: object
                        required:
                            - id
                            - type
                            - tag
                            - label
                            - schema
                            - owner
                            - document
                        properties:
                            id:
                                type: string
                            type:
                                type: string
                            tag:
                                type: string
                            label:
                                type: string
                            schema:
                                type: string
                            owner:
                                type: string
                            document:
                                type: object
                userMap:
                    type: array
                    items:
                        type: object
                        required:
                            - did
                            - username
                        properties:
                            did:
                                type: string
                            username:
                                type: string
        VerifiablePresentation:
            type: object
            required:
                - hash
                - id
                - policyId
                - signature
                - status
                - tag
                - type
                - updateDate
                - createDate
                - owner
                - document
            properties:
                hash:
                    type: string
                id:
                    type: string
                policyId:
                    type: string
                signature:
                    type: string
                status:
                    type: string
                tag:
                    type: string
                type:
                    type: string
                updateDate:
                    type: string
                createDate:
                    type: string
                owner:
                    type: string
                document:
                    type: object
...
        Error:
            type: object
            required:
                - code
                - message
            properties:
                code:
                    type: number
                message:
                    type: string
...
paths:
...
    /trustchains:
        get:
            tags:
                - trustchains
            description: Requests all VP documents. Only users with the Auditor role are allowed to make the request.
            security:
                - bearerAuth: []
            summary: Returns a list of all VP documents.
            responses:
                200:
                    description: Successful operation.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/VerifiablePresentation'
                401:
                    description: Unauthorized.
                403:
                    description: Forbidden.
                500:
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /trustchains/{hash}:
        get:
            tags:
                - trustchains
            description: Builds and returns a trustchain, from the VP to the root VC document. Only users with the Auditor role are allowed to make the request.
            parameters:
                - in: path
                  name: hash
                  schema:
                      type: string
                  required: true
                  description: Hash or ID of a VP document.
            security:
                - bearerAuth: []
            summary: Returns a trustchain for a VP document.
            responses:
                200:
                    description: Successful operation.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TrustChains'
                401:
                    description: Unauthorized.
                403:
                    description: Forbidden.
                500:
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...